set up my folder structures
