+++ controllers/products.js
@@ -1,0 +1,5 @@
+var models = require('../models');
+
+module.exports = {
+
+};
