+++ models/index.js
@@ -1,0 +1,1 @@
+module.exports.products = require('./products.js');
