+++ models/products.js
@@ -1,0 +1,5 @@
+var db = require('../database/db.js');
+
+module.exports = {
+
+};
