+++ routes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const controllers = require('./controllers');
+
+const router = express.Router();
+
+// router.get();
+
+module.exports = router;
